core: dts: phycore-stm32mp15: fix memory node
The DDR size in "stm32mp15xx-phycore-som.dtsi" is not correct as two variants
are available for standard phyCORE-STM32MP15: 1GiB or 512MB.
So we must set the "memory" node in the ".dts" file instead.
For "stm32mp157c-phycore-som-emmc-1gib.dts", it is 1GiB DDR.

This modification is necessary when using optee.

Signed-off-by: Christophe Parant <c.parant@phytec.fr>